moved all tablet code to Tablet namespace
