+++ Assets/Scripts/Networking/Tablet.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6c744df940cc44af99934fc90e79568a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
